Added minor bug fixes (some expressions starting with ^ were not being found), plus a few warning suppressions.
git-svn-id: http://svn.boost.org/svn/boost/trunk@18679 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/boost/regex/v4/match_results.hpp
@@ -103,7 +103,7 @@ public:
    {
       sub += 2;
       string_type result;
-      if(sub < m_subs.size() && (sub > 0))
+      if(sub < (int)m_subs.size() && (sub > 0))
       {
          const sub_match<RandomAccessIterator>& s = m_subs[sub];
          if(s.matched)
